requirements.txt
djangorestframework
pygments